@@ -1,20 +1,17 @@
 name: simple_code
 description: Make Development Faster and Easier
-version: 5.1.0
+version: 6.0.0
 homepage: https://github.com/thize/simple_code
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  logger: ^0.9.4
-  simple_animations: ^2.4.2
-  time: ^1.4.0
-  url_launcher: ^5.7.10
+  logger: ^1.0.0-nullsafety.0
+  simple_animations: ^3.0.0-nullsafety.4
+  auto_size_text: ^3.0.0-nullsafety.0
 
 dev_dependencies:
-  lint: ^1.2.0
-
-flutter:
+  lint: ^1.5.2
