@@ -1,13 +1,21 @@
 part of simple_code;
 
 class ExpandedText extends StatefulWidget {
+  /// Creates a [ExpandedText] widget.
+  ///
+  /// If the [style] argument is null, the text will use the style from the
+  /// closest enclosing [DefaultTextStyle].
   const ExpandedText(
-    this.data, {
-    Key key,
+    String this.data, {
+    Key? key,
+    this.textKey,
     this.style,
     this.strutStyle,
     this.minFontSize = 1,
     this.maxFontSize = double.maxFinite,
+    this.stepGranularity = 1,
+    this.presetFontSizes,
+    this.group,
     this.textAlign,
     this.textDirection,
     this.locale,
@@ -18,56 +26,168 @@ class ExpandedText extends StatefulWidget {
     this.textScaleFactor,
     this.maxLines,
     this.semanticsLabel,
-  })  : assert(data != null,
-            'A non-null String must be provided to a ExpandedText widget.'),
-        textSpan = null,
+  })  : textSpan = null,
         super(key: key);
-  const ExpandedText.rich(
-    this.textSpan, {
-    Key key,
-    this.style,
-    this.strutStyle,
-    this.minFontSize = 1,
-    this.maxFontSize = double.maxFinite,
-    this.textAlign,
-    this.textDirection,
-    this.locale,
-    this.softWrap,
-    this.wrapWords = true,
-    this.overflow,
-    this.overflowReplacement,
-    this.textScaleFactor,
-    this.maxLines,
-    this.semanticsLabel,
-  })  : assert(textSpan != null,
-            // ignore: lines_longer_than_80_chars
-            'A non-null TextSpan must be provided to a ExpandedText.rich widget.'),
-        data = null,
-        super(key: key);
-  final String data;
-  final TextSpan textSpan;
-  final TextStyle style;
-  final StrutStyle strutStyle;
+
+  /// Sets the key for the resulting [Text] widget.
+  ///
+  /// This allows you to find the actual `Text` widget built by `AutoSizeText`.
+  final Key? textKey;
+
+  /// The text to display.
+  ///
+  /// This will be null if a [textSpan] is provided instead.
+  final String? data;
+
+  /// The text to display as a [TextSpan].
+  ///
+  /// This will be null if [data] is provided instead.
+  final TextSpan? textSpan;
+
+  /// If non-null, the style to use for this text.
+  ///
+  /// If the style's "inherit" property is true, the style will be merged with
+  /// the closest enclosing [DefaultTextStyle]. Otherwise, the style will
+  /// replace the closest enclosing [DefaultTextStyle].
+  final TextStyle? style;
+
+  /// The strut style to use. Strut style defines the strut, which sets minimum
+  /// vertical layout metrics.
+  ///
+  /// Omitting or providing null will disable strut.
+  ///
+  /// Omitting or providing null for any properties of [StrutStyle] will result
+  /// in default values being used. It is highly recommended to at least specify
+  /// a font size.
+  ///
+  /// See [StrutStyle] for details.
+  final StrutStyle? strutStyle;
+
+  /// The minimum text size constraint to be used when auto-sizing text.
+  ///
+  /// Is being ignored if [presetFontSizes] is set.
   final double minFontSize;
+
+  /// The maximum text size constraint to be used when auto-sizing text.
+  ///
+  /// Is being ignored if [presetFontSizes] is set.
   final double maxFontSize;
-  final TextAlign textAlign;
-  final TextDirection textDirection;
-  final Locale locale;
-  final bool softWrap;
+
+  /// The step size in which the font size is being adapted to constraints.
+  ///
+  /// The Text scales uniformly in a range between [minFontSize] and
+  /// [maxFontSize].
+  /// Each increment occurs as per the step size set in stepGranularity.
+  ///
+  /// Most of the time you don't want a stepGranularity below 1.0.
+  ///
+  /// Is being ignored if [presetFontSizes] is set.
+  final double stepGranularity;
+
+  /// Predefines all the possible font sizes.
+  ///
+  /// **Important:** PresetFontSizes have to be in descending order.
+  final List<double>? presetFontSizes;
+
+  /// Synchronizes the size of multiple [AutoSizeText]s.
+  ///
+  /// If you want multiple [AutoSizeText]s to have the same text size, give all
+  /// of them the same [AutoSizeGroup] instance. All of them will have the
+  /// size of the smallest [AutoSizeText]
+  final AutoSizeGroup? group;
+
+  /// How the text should be aligned horizontally.
+  final TextAlign? textAlign;
+
+  /// The directionality of the text.
+  ///
+  /// This decides how [textAlign] values like [TextAlign.start] and
+  /// [TextAlign.end] are interpreted.
+  ///
+  /// This is also used to disambiguate how to render bidirectional text. For
+  /// example, if the [data] is an English phrase followed by a Hebrew phrase,
+  /// in a [TextDirection.ltr] context the English phrase will be on the left
+  /// and the Hebrew phrase to its right, while in a [TextDirection.rtl]
+  /// context, the English phrase will be on the right and the Hebrew phrase on
+  /// its left.
+  ///
+  /// Defaults to the ambient [Directionality], if any.
+  final TextDirection? textDirection;
+
+  /// Used to select a font when the same Unicode character can
+  /// be rendered differently, depending on the locale.
+  ///
+  /// It's rarely necessary to set this property. By default its value
+  /// is inherited from the enclosing app with `Localizations.localeOf(context)`.
+  final Locale? locale;
+
+  /// Whether the text should break at soft line breaks.
+  ///
+  /// If false, the glyphs in the text will be positioned as if there was
+  /// unlimited horizontal space.
+  final bool? softWrap;
+
+  /// Whether words which don't fit in one line should be wrapped.
+  ///
+  /// If false, the fontSize is lowered as far as possible until all words fit
+  /// into a single line.
   final bool wrapWords;
-  final TextOverflow overflow;
-  final Widget overflowReplacement;
-  final double textScaleFactor;
-  final int maxLines;
-  final String semanticsLabel;
+
+  /// How visual overflow should be handled.
+  ///
+  /// Defaults to retrieving the value from the nearest [DefaultTextStyle] ancestor.
+  final TextOverflow? overflow;
+
+  /// If the text is overflowing and does not fit its bounds, this widget is
+  /// displayed instead.
+  final Widget? overflowReplacement;
+
+  /// The number of font pixels for each logical pixel.
+  ///
+  /// For example, if the text scale factor is 1.5, text will be 50% larger than
+  /// the specified font size.
+  ///
+  /// This property also affects [minFontSize], [maxFontSize] and [presetFontSizes].
+  ///
+  /// The value given to the constructor as textScaleFactor. If null, will
+  /// use the [MediaQueryData.textScaleFactor] obtained from the ambient
+  /// [MediaQuery], or 1.0 if there is no [MediaQuery] in scope.
+  final double? textScaleFactor;
+
+  /// An optional maximum number of lines for the text to span, wrapping if necessary.
+  /// If the text exceeds the given number of lines, it will be resized according
+  /// to the specified bounds and if necessary truncated according to [overflow].
+  ///
+  /// If this is 1, text will not wrap. Otherwise, text will be wrapped at the
+  /// edge of the box.
+  ///
+  /// If this is null, but there is an ambient [DefaultTextStyle] that specifies
+  /// an explicit number for its [DefaultTextStyle.maxLines], then the
+  /// [DefaultTextStyle] value will take precedence. You can use a [RichText]
+  /// widget directly to entirely override the [DefaultTextStyle].
+  final int? maxLines;
+
+  /// An alternative semantics label for this text.
+  ///
+  /// If present, the semantics of this widget will contain this value instead
+  /// of the actual text. This will overwrite any of the semantics labels applied
+  /// directly to the [TextSpan]s.
+  ///
+  /// This is useful for replacing abbreviations or shorthands with the full
+  /// text value:
+  ///
+  /// ```dart
+  /// AutoSizeText(r'$$', semanticsLabel: 'Double dollars')
+  /// ```
+  final String? semanticsLabel;
 
   @override
   ExpandedTextState createState() => ExpandedTextState();
 }
 
 class ExpandedTextState extends State<ExpandedText> {
-  double maxFontSize;
-  double minFontSize;
+  late double maxFontSize;
+  late double minFontSize;
   final double stepGranularity = 1;
   @override
   void initState() {
@@ -91,171 +211,51 @@ class ExpandedTextState extends State<ExpandedText> {
 
   @override
   Widget build(BuildContext context) {
-    return LayoutBuilder(builder: (context, size) {
-      final DefaultTextStyle defaultTextStyle = DefaultTextStyle.of(context);
-
-      TextStyle style = widget.style;
-      if (widget.style == null || widget.style.inherit) {
-        style = defaultTextStyle.style.merge(widget.style);
-      }
-
-      style = style.copyWith(fontSize: maxFontSize);
-
-      final maxLines = widget.maxLines ?? defaultTextStyle.maxLines;
-
-      _sanityCheck(style, maxLines);
-
-      final result = _calculateFontSize(size, style, maxLines);
-      final double fontSize = result[0] as double;
-      final bool textFits = result[1] as bool;
-
-      Widget text;
-
-      text = _buildText(fontSize, style, maxLines);
-
-      if (widget.overflowReplacement != null && !textFits) {
-        return widget.overflowReplacement;
-      } else {
-        return text;
-      }
-    });
-  }
-
-  void _sanityCheck(TextStyle style, int maxLines) {
-    assert(widget.overflow == null || widget.overflowReplacement == null,
-        'Either overflow or overflowReplacement have to be null.');
-    assert(maxLines == null || maxLines > 0,
-        'MaxLines has to be grater than or equal to 1.');
-
-    assert(
-        stepGranularity >= 0.1,
-        // ignore: lines_longer_than_80_chars
-        'StepGranularity has to be greater than or equal to 0.1. It is not a good idea to resize the font with a higher accuracy.');
-    assert(
-        minFontSize >= 0, 'MinFontSize has to be greater than or equal to 0.');
-    assert(maxFontSize > 0, 'MaxFontSize has to be greater than 0.');
-    assert(minFontSize <= maxFontSize,
-        'MinFontSize has to be smaller or equal than maxFontSize.');
-    assert(minFontSize / stepGranularity % 1 == 0,
-        'MinFontSize has to be multiples of stepGranularity.');
-    if (maxFontSize != double.maxFinite) {
-      assert(maxFontSize / stepGranularity % 1 == 0,
-          'MaxFontSize has to be multiples of stepGranularity.');
-    }
-    assert(style.fontSize / stepGranularity % 1 == 0,
-        'FontSize has to be multiples of stepGranularity.');
-  }
-
-  List _calculateFontSize(BoxConstraints size, TextStyle style, int maxLines) {
-    final span = TextSpan(
-      style: widget.textSpan?.style ?? style,
-      text: widget.textSpan?.text ?? widget.data,
-      children: widget.textSpan?.children,
-      recognizer: widget.textSpan?.recognizer,
-    );
-
-    int left;
-    int right;
-
-    left = (minFontSize / stepGranularity).round();
-    final initialFontSize = style.fontSize.clamp(minFontSize, maxFontSize);
-    right = (initialFontSize / stepGranularity).round();
-
-    final userScale =
-        widget.textScaleFactor ?? MediaQuery.textScaleFactorOf(context);
-
-    bool _testValue(int value) {
-      double scale;
-      scale = value * userScale * stepGranularity / style.fontSize;
-
-      return _checkTextFits(span, scale, maxLines, size);
-    }
-
-    bool lastValueFits = false;
-    if (_testValue(right)) {
-      lastValueFits = true;
-    } else {
-      right -= 1;
-      while (left <= right) {
-        final mid = (left + (right - left) / 2).toInt();
-        if (_testValue(mid)) {
-          left = mid + 1;
-          lastValueFits = true;
-        } else {
-          right = mid - 1;
-        }
-      }
-
-      if (!lastValueFits) {
-        right += 1;
-      }
-    }
-
-    double fontSize;
-    fontSize = right * userScale * stepGranularity;
-
-    return [fontSize, lastValueFits];
-  }
-
-  bool _checkTextFits(
-      TextSpan text, double scale, int maxLines, BoxConstraints constraints) {
-    if (!widget.wrapWords) {
-      final wordCount = text.toPlainText().split(RegExp('\\s+')).length;
-      if (maxLines == null) {
-        // ignore: parameter_assignments
-        maxLines = wordCount;
-      } else {
-        // ignore: parameter_assignments
-        maxLines = maxLines.clamp(1, wordCount).toInt();
-      }
-    }
-
-    final tp = TextPainter(
-      text: text,
-      textAlign: widget.textAlign ?? TextAlign.left,
-      textDirection: widget.textDirection ?? TextDirection.ltr,
-      textScaleFactor: scale ?? 1,
-      maxLines: maxLines,
-      locale: widget.locale,
-      strutStyle: widget.strutStyle,
-    );
-
-    tp.layout(maxWidth: constraints.maxWidth);
-
-    return !(tp.didExceedMaxLines ||
-        tp.height > constraints.maxHeight ||
-        tp.width > constraints.maxWidth);
-  }
-
-  Widget _buildText(double fontSize, TextStyle style, int maxLines) {
+    final TextStyle textStyle =
+        widget.style ?? DefaultTextStyle.of(context).style;
     if (widget.data != null) {
-      return Text(
-        widget.data,
-        style: style.copyWith(fontSize: fontSize),
+      return AutoSizeText(
+        widget.data!,
+        style: textStyle.copyWith(fontSize: maxFontSize),
         strutStyle: widget.strutStyle,
         textAlign: widget.textAlign,
         textDirection: widget.textDirection,
         locale: widget.locale,
         softWrap: widget.softWrap,
         overflow: widget.overflow,
-        textScaleFactor: 1,
-        maxLines: maxLines,
+        textScaleFactor: widget.textScaleFactor,
+        maxLines: widget.maxLines,
         semanticsLabel: widget.semanticsLabel,
-      );
-    } else {
-      return Text.rich(
-        widget.textSpan,
-        style: style,
-        strutStyle: widget.strutStyle,
-        textAlign: widget.textAlign,
-        textDirection: widget.textDirection,
-        locale: widget.locale,
-        softWrap: widget.softWrap,
-        overflow: widget.overflow,
-        textScaleFactor: fontSize / style.fontSize,
-        maxLines: maxLines,
-        semanticsLabel: widget.semanticsLabel,
+        group: widget.group,
+        maxFontSize: maxFontSize,
+        minFontSize: minFontSize,
+        overflowReplacement: widget.overflowReplacement,
+        presetFontSizes: widget.presetFontSizes,
+        stepGranularity: widget.stepGranularity,
+        textKey: widget.textKey,
+        wrapWords: widget.wrapWords,
       );
     }
+    return AutoSizeText.rich(
+      widget.textSpan!,
+      style: textStyle.copyWith(fontSize: maxFontSize),
+      strutStyle: widget.strutStyle,
+      textAlign: widget.textAlign,
+      textDirection: widget.textDirection,
+      locale: widget.locale,
+      softWrap: widget.softWrap,
+      overflow: widget.overflow,
+      textScaleFactor: widget.textScaleFactor,
+      maxLines: widget.maxLines,
+      semanticsLabel: widget.semanticsLabel,
+      group: widget.group,
+      maxFontSize: maxFontSize,
+      minFontSize: minFontSize,
+      overflowReplacement: widget.overflowReplacement,
+      presetFontSizes: widget.presetFontSizes,
+      stepGranularity: widget.stepGranularity,
+      textKey: widget.textKey,
+      wrapWords: widget.wrapWords,
+    );
   }
 }
