@@ -10,9 +10,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/material.dart' as material;
 import 'package:flutter/services.dart';
 import 'package:logger/logger.dart';
+import 'package:auto_size_text/auto_size_text.dart';
 import 'package:simple_animations/simple_animations.dart';
-import 'package:time/time.dart';
-import 'package:url_launcher/url_launcher.dart';
 
 part 'src/extensions/bool_extension.dart';
 part 'src/extensions/color_extension.dart';
@@ -25,9 +24,6 @@ part 'src/simple_code.dart';
 part 'src/simple_navigator/navigator.dart';
 part 'src/simple_navigator/route.dart';
 part 'src/simple_utils.dart';
-part 'src/simple_widgets/animated_button.dart';
-part 'src/simple_widgets/animated_tween.dart';
 part 'src/simple_widgets/animation.dart';
-part 'src/simple_widgets/expanded_icon.dart';
 part 'src/simple_widgets/expanded_text.dart';
 part 'src/simple_widgets/tap_animation.dart';
